MAJOR: Initial version of the Explainer
This commit separates the Explainer software from the Reporter. As the
software is, effectively, a fork of the Reporter, we retain all the
previous git commit history.

Major changes include renaming a large proportion of the files, changing
the Fact data structure to better suit the Explanation task, and
significant reduction in processing complexity. For example the
Aggregation task is completely removed from the pipeline for now.